Fix errors with functionality
