+++ app/javascript/redux/store.js
@@ -1,8 +1,9 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { greetingSlice } from './reducer/greetingSlice';
+// Path: app\javascript\redux\store.js
+import { configureStore } from "@reduxjs/toolkit";
+import greetingSlice from "./greetingSlice";
 
 export default configureStore({
   reducer: {
-    greeting: greetingSlice
-  }
+    greeting: greetingSlice,
+  },
 });
