+++ app/models/greeting.rb
