+++ config/routes.rb
@@ -1,7 +1,9 @@
 Rails.application.routes.draw do
-  get 'root/index'
-  # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
-
-  # Defines the root path route ("/")
   root "root#index"
+
+  namespace :api do
+    namespace :v1 do
+      resources :greetings, only: [:index]
+    end
+  end
 end
