+++ package.json
@@ -2,6 +2,7 @@
   "name": "hello-rails-react",
   "private": "true",
   "dependencies": {
+    "@babel/preset-react": "^7.18.6",
     "@hotwired/stimulus": "^3.1.1",
     "@hotwired/turbo-rails": "^7.2.4",
     "@reduxjs/toolkit": "^1.9.0",
@@ -19,7 +20,6 @@
   "devDependencies": {
     "@babel/core": "^7.20.2",
     "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
     "babel-loader": "^9.1.0",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
